requirements.txt
altgraph==0.15
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
decorator==4.3.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
future==0.16.0
gunicorn==19.8.1
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.1
macholib==1.9
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.3
parso==0.2.1
pefile==2017.11.5
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyInstaller==3.3.1
python-dateutil==2.7.3
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
simplegeneric==0.8.1
six==1.11.0
tldextract==2.2.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
